+++ lib/src/simpleCrop.dart
@@ -2,16 +2,25 @@ part of image_crop;
 
 const _kCropOverlayActiveOpacity = 0.3;
 const _kCropOverlayInactiveOpacity = 0.7;
-const _kCropHandleSize = 10.0;
 
 enum _CropAction { none, moving, scaling }
+
+enum ChipShape {
+  /// Crop rectangle
+  rect,
+
+  /// Crop circle
+  circle,
+}
 
 class ImgCrop extends StatefulWidget {
   final ImageProvider image;
   final double maximumScale;
   final ImageErrorListener onImageError;
-  final double chipRadius; // 裁剪半径
-  final String chipShape; // 裁剪区域形状
+  final double chipRadius;
+  final double chipRatio;
+  final ChipShape chipShape;
+  final double handleSize;
   final Color borderColor;
   const ImgCrop(
       {Key key,
@@ -19,10 +28,13 @@ class ImgCrop extends StatefulWidget {
       this.maximumScale: 2.0,
       this.onImageError,
       this.chipRadius = 150,
-      this.chipShape = 'circle',
+      this.chipRatio = 1.0,
+      this.chipShape = ChipShape.circle,
+      this.handleSize = 10.0,
       this.borderColor})
       : assert(image != null),
         assert(maximumScale != null),
+        assert(handleSize != null && handleSize >= 0.0),
         super(key: key);
 
   ImgCrop.file(File file,
@@ -31,23 +43,26 @@ class ImgCrop extends StatefulWidget {
       this.maximumScale: 2.0,
       this.onImageError,
       this.chipRadius = 150,
-      this.chipShape = 'circle',
+      this.chipRatio = 1.0,
+      this.chipShape = ChipShape.circle,
+      this.handleSize = 10.0,
       this.borderColor})
       : image = FileImage(file, scale: scale),
         assert(maximumScale != null),
         super(key: key);
 
-  ImgCrop.asset(
-    String assetName, {
-    Key key,
-    AssetBundle bundle,
-    String package,
-    this.chipRadius = 150,
-    this.maximumScale: 2.0,
-    this.onImageError,
-    this.chipShape = 'circle',
-    this.borderColor,
-  })  : image = AssetImage(assetName, bundle: bundle, package: package),
+  ImgCrop.asset(String assetName,
+      {Key key,
+      AssetBundle bundle,
+      String package,
+      this.chipRadius = 150,
+      this.maximumScale: 2.0,
+      this.onImageError,
+      this.chipRatio = 1.0,
+      this.chipShape = ChipShape.circle,
+      this.handleSize = 10.0,
+      this.borderColor})
+      : image = AssetImage(assetName, bundle: bundle, package: package),
         assert(maximumScale != null),
         super(key: key);
 
@@ -132,19 +147,21 @@ class ImgCropState extends State<ImgCrop> with TickerProviderStateMixin, Drag {
     _activate(1.0);
   }
 
-  Future<File> cropCompleted(File file, {int pictureQuality}) async {
-    final options = await ImageCrop.getImageOptions(file: file);
-    debugPrint('image width: ${options.width}, height: ${options.height}');
+  Future<File> cropCompleted(File file, {int preferredSize}) async {
+    // final options = await ImageCrop.getImageOptions(file: file);
+    // debugPrint(
+    //     'image width: ${options.width}, height: ${options.height}  $scale');
+
     final sampleFile = await ImageCrop.sampleImage(
       file: file,
-      preferredWidth: (pictureQuality / scale).round(),
-      preferredHeight: (pictureQuality / scale).round(),
+      preferredSize: (preferredSize / scale).round(),
     );
 
     final croppedFile = await ImageCrop.cropImage(
       file: sampleFile,
       area: area,
     );
+
     return croppedFile;
   }
 
@@ -172,13 +189,15 @@ class ImgCropState extends State<ImgCrop> with TickerProviderStateMixin, Drag {
         child: CustomPaint(
           painter: _CropPainter(
               borderColor: widget.borderColor,
-              image: _image,
-              ratio: _ratio,
-              view: _view,
-              area: _area,
-              scale: _scale,
-              active: _activeController.value,
-              chipShape: widget.chipShape),
+            image: _image,
+            ratio: _ratio,
+            view: _view,
+            area: _area,
+            scale: _scale,
+            active: _activeController.value,
+            chipShape: widget.chipShape,
+            handleSize: widget.handleSize,
+          ),
         ),
       ),
     );
@@ -195,7 +214,7 @@ class ImgCropState extends State<ImgCrop> with TickerProviderStateMixin, Drag {
   // NOTE: 区域性缩小 总区域 - 10 * 10 区域
   Size get _boundaries {
     return _surfaceKey.currentContext.size -
-        Offset(_kCropHandleSize, _kCropHandleSize);
+        Offset(widget.handleSize, widget.handleSize);
   }
 
   void _settleAnimationChanged() {
@@ -217,13 +236,15 @@ class ImgCropState extends State<ImgCrop> with TickerProviderStateMixin, Drag {
     }
 
     final _deviceWidth =
-        MediaQuery.of(context).size.width - (2 * _kCropHandleSize);
+        MediaQuery.of(context).size.width - (2 * widget.handleSize);
     final _areaOffset = (_deviceWidth - (widget.chipRadius * 2));
     final _areaOffsetRadio = _areaOffset / _deviceWidth;
     final width = 1.0 - _areaOffsetRadio;
+    final height = (imageWidth * viewWidth * width) /
+        (imageHeight *
+            viewHeight *
+            (widget.chipShape == ChipShape.rect ? widget.chipRatio : 1.0));
 
-    final height =
-        (imageWidth * viewWidth * width) / (imageHeight * viewHeight * 1.0);
     return Rect.fromLTWH((1.0 - width) / 2, (1.0 - height) / 2, width, height);
   }
 
@@ -370,7 +391,8 @@ class _CropPainter extends CustomPainter {
   final Rect area;
   final double scale;
   final double active;
-  final String chipShape;
+  final ChipShape chipShape;
+  final double handleSize;
   Color borderColor = Colors.white;
 
   _CropPainter(
@@ -381,6 +403,7 @@ class _CropPainter extends CustomPainter {
       this.scale,
       this.active,
       this.chipShape,
+      this.handleSize,
       this.borderColor});
 
   @override
@@ -395,10 +418,10 @@ class _CropPainter extends CustomPainter {
 
   currentRact(size) {
     return Rect.fromLTWH(
-      _kCropHandleSize / 2,
-      _kCropHandleSize / 2,
-      size.width - _kCropHandleSize,
-      size.height - _kCropHandleSize,
+      handleSize / 2,
+      handleSize / 2,
+      size.width - handleSize,
+      size.height - handleSize,
     );
   }
 
@@ -451,13 +474,8 @@ class _CropPainter extends CustomPainter {
     final _path1 = Path()
       ..addRect(Rect.fromLTRB(0.0, 0.0, rect.width, rect.height));
     Path _path2;
-    if (chipShape == 'rect') {
-      _path2 = Path()..addRRect(RRect.fromLTRBR(
-          boundaries.left,
-          boundaries.top,
-          boundaries.right,
-          boundaries.bottom,
-          Radius.circular(10)));
+    if (chipShape == ChipShape.rect) {
+      _path2 = Path()..addRect(boundaries);
     } else {
       _path2 = Path()
         ..addRRect(RRect.fromLTRBR(
@@ -475,10 +493,10 @@ class _CropPainter extends CustomPainter {
       ..color = borderColor
       ..strokeWidth = 2
       ..style = PaintingStyle.stroke;
-    if (chipShape == 'rect') {
-      canvas.drawRRect(
-          RRect.fromLTRBR(boundaries.left - 1, boundaries.top - 1,
-              boundaries.right + 1, boundaries.bottom + 1,Radius.circular(10)),
+    if (chipShape == ChipShape.rect) {
+      canvas.drawRect(
+          Rect.fromLTRB(boundaries.left - 1, boundaries.top - 1,
+              boundaries.right + 1, boundaries.bottom + 1),
           paint);
     } else {
       canvas.drawRRect(