+++ CHANGELOG.md
@@ -17,3 +17,13 @@ update README
 ## 0.1.2
 
 update api findAncestorStateOfType
+
+## 0.1.3
+
+update README
+
+## 1.0.0
+
+Add API `chipRatio` to control the rectangle aspect ratio
+
+Add API `handleSize` to control the chip handle area